requirements.txt
pyyaml==6.0.1
requests==2.31.0
watchdog==3.0.0
